// admin/src/api/notification/Notification.ts
import { RegistrationWhereUniqueInput } from "../registration/RegistrationWhereUniqueInput";

export type Notification = {
  createdAt: Date;
  disabled: boolean | null;
  email: string;
  fax: string | null;
  id: string;
  phone: string;
  registrationId?: RegistrationWhereUniqueInput | null;
  sendFax: boolean | null;
  sendSms: boolean | null;
  updatedAt: Date;
};